LIBFCREPO-1189. Added plastron-models as a separate distribution package.
Contains these modules:
- plastron.files
- plastron.models
- plastron.namespaces
- plastron.rdf
- plastron.serializers
- plastron.validation

https://umd-dit.atlassian.net/browse/LIBFCREPO-1189
