+++ plastron-legacy/src/plastron/handlers/mapped-csv.py
@@ -5,7 +5,7 @@ import yaml
 from rdflib import Literal, URIRef
 from rdflib.util import from_n3
 from plastron import pcdm, namespaces, rdf
-from plastron.files import LocalFileSource, RemoteFileSource
+from plastron.files.__init__ import LocalFileSource, RemoteFileSource
 from plastron.exceptions import ConfigError
 from collections import OrderedDict
