+++ plastron-legacy/src/plastron/jobs.py
@@ -16,8 +16,8 @@ from rdflib.util import from_n3
 import plastron.models
 from plastron.exceptions import DataReadException, FailureException
 from plastron.namespaces import get_manager
-from plastron.rdf import RDFDataProperty, Resource
-from plastron.serializers import CSVSerializer
+from plastron.rdf.rdf import RDFDataProperty, Resource
+from plastron.serializers.__init__ import CSVSerializer
 from plastron.util import ItemLog, datetimestamp
