--- plastron-legacy/src/plastron/logging.py
@@ -1,58 +1,0 @@
-import json
-import logging
-
-DEFAULT_LOGGING_OPTIONS = {
-    'version': 1,
-    'formatters': {
-        'full': {
-            'format': '%(levelname)s|%(asctime)s|%(threadName)s|%(name)s|%(message)s'
-        },
-        'messageonly': {
-            'format': '%(message)s'
-        }
-    },
-    'handlers': {
-        'console': {
-            'class': 'logging.StreamHandler',
-            'level': 'INFO',
-            'formatter': 'messageonly',
-            'stream': 'ext://sys.stderr'
-        },
-        'file': {
-            'class': 'logging.FileHandler',
-            'level': 'DEBUG',
-            'formatter': 'full'
-        }
-    },
-    'loggers': {
-        '__main__': {
-            'level': 'DEBUG',
-            'handlers': ['console', 'file'],
-            'propagate': False
-        },
-        'plastron': {
-            'level': 'DEBUG',
-            'handlers': ['console', 'file'],
-            'propagate': False
-        },
-        # suppress logging output from paramiko by default
-        'paramiko': {
-            'propagate': False
-        }
-    },
-    'root': {
-        'level': 'DEBUG'
-    }
-}
-
-
-class STOMPHandler(logging.Handler):
-    def __init__(self, level=logging.NOTSET, connection=None, destination=None):
-        super().__init__(level)
-        self.destination = destination
-        self.connection = connection
-
-    def emit(self, record):
-        # if no broker is set, just be silent
-        if self.connection is not None and self.destination is not None:
-            self.connection.send(self.destination, headers=record.msg.headers, body=record.msg.body)
