--- plastron-legacy/src/plastron/models/__init__.py
@@ -1,4 +1,0 @@
-from .letter import *
-from .newspaper import *
-from .poster import *
-from .umd import *
