+++ plastron-legacy/src/plastron/util.py
@@ -22,6 +22,50 @@ from plastron.client import Client, TransactionClient
 from plastron.exceptions import FailureException, RESTAPIException
 from plastron.namespaces import dcterms
 
+DEFAULT_LOGGING_OPTIONS = {
+    'version': 1,
+    'formatters': {
+        'full': {
+            'format': '%(levelname)s|%(asctime)s|%(threadName)s|%(name)s|%(message)s'
+        },
+        'messageonly': {
+            'format': '%(message)s'
+        }
+    },
+    'handlers': {
+        'console': {
+            'class': 'logging.StreamHandler',
+            'level': 'INFO',
+            'formatter': 'messageonly',
+            'stream': 'ext://sys.stderr'
+        },
+        'file': {
+            'class': 'logging.FileHandler',
+            'level': 'DEBUG',
+            'formatter': 'full'
+        }
+    },
+    'loggers': {
+        '__main__': {
+            'level': 'DEBUG',
+            'handlers': ['console', 'file'],
+            'propagate': False
+        },
+        'plastron': {
+            'level': 'DEBUG',
+            'handlers': ['console', 'file'],
+            'propagate': False
+        },
+        # suppress logging output from paramiko by default
+        'paramiko': {
+            'propagate': False
+        }
+    },
+    'root': {
+        'level': 'DEBUG'
+    }
+}
+
 logger = logging.getLogger(__name__)
