+++ plastron-legacy/tests/test_embedded_objects.py
@@ -1,5 +1,4 @@
 import csv
-import io
 from unittest.mock import Mock
 
 import pytest
