+++ plastron-models/pyproject.toml
@@ -1,0 +1,16 @@
+[project]
+name = "plastron-models"
+version = "4.0.0-dev"
+requires-python = ">= 3.8"
+dependencies = [
+    "requests",
+    "paramiko",
+    "rdflib",
+]
+
+[project.optional-dependencies]
+test = [
+    "freezegun",
+    "pytest",
+    "pytest-cov",
+]
