+++ plastron-models/src/plastron/models/letter.py
@@ -1,9 +1,9 @@
-from plastron import rdf, pcdm
-from plastron.authority import LabeledThing
-from plastron.namespaces import bibo, dc, dcmitype, dcterms, edm, geo, rel, skos
-from plastron.validation import is_edtf_formatted, is_handle
 from rdflib import Namespace
 
+from plastron.validation import is_edtf_formatted, is_handle
+from plastron.namespaces import bibo, dc, dcmitype, dcterms, edm, geo, rel, skos
+from plastron.rdf import pcdm, rdf
+from plastron.rdf.authority import LabeledThing
 
 umdtype = Namespace('http://vocab.lib.umd.edu/datatype#')
