+++ plastron-models/src/plastron/models/poster.py
@@ -1,8 +1,8 @@
-from plastron import pcdm, rdf
-from plastron.namespaces import dcterms, dc, edm, bibo, geo
-from plastron.validation import is_edtf_formatted, is_handle
 from rdflib import Namespace
 
+from plastron.rdf import pcdm, rdf
+from plastron.validation import is_edtf_formatted, is_handle
+from plastron.namespaces import dcterms, dc, edm, bibo, geo
 
 umdtype = Namespace('http://vocab.lib.umd.edu/datatype#')
