+++ plastron-models/src/plastron/validation/vocabularies/dcmitype.ttl
