+++ plastron-models/tests/test_files/sample.zip
